fix: split OpenAPI docs alignment tests to satisfy backend-size limit
